startup logo

+++ include/lib/types.h
@@ -4,7 +4,8 @@
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
+
 typedef unsigned long size_t;
 
 typedef signed char int8_t;
